requirements.txt
pymupdf
pillow 
pytesseract
tesseract
langchain 
langchain-community